edit publish

+++ src/main/java/com/lingerlin/community/community/model/Discussion.java
@@ -3,15 +3,15 @@ package com.lingerlin.community.community.model;
 
 public class Discussion {
 
-  private long id;
+  private Integer id;
   private String title;
   private String description;
-  private long gmtCreate;
-  private long gmtModified;
-  private long creator;
-  private long commentCount;
-  private long viewCount;
-  private long likeCount;
+  private Long gmtCreate;
+  private Long gmtModified;
+  private Integer creator;
+  private Integer commentCount;
+  private Integer viewCount;
+  private Integer likeCount;
   private String tag;
 
 
@@ -19,7 +19,7 @@ public class Discussion {
     return id;
   }
 
-  public void setId(long id) {
+  public void setId(Integer id) {
     this.id = id;
   }
 
@@ -42,56 +42,56 @@ public class Discussion {
   }
 
 
-  public long getGmtCreate() {
+  public Long getGmtCreate() {
     return gmtCreate;
   }
 
-  public void setGmtCreate(long gmtCreate) {
+  public void setGmtCreate(Long gmtCreate) {
     this.gmtCreate = gmtCreate;
   }
 
 
-  public long getGmtModified() {
+  public Long getGmtModified() {
     return gmtModified;
   }
 
-  public void setGmtModified(long gmtModified) {
+  public void setGmtModified(Long gmtModified) {
     this.gmtModified = gmtModified;
   }
 
 
-  public long getCreator() {
+  public Integer getCreator() {
     return creator;
   }
 
-  public void setCreator(long creator) {
+  public void setCreator(Integer creator) {
     this.creator = creator;
   }
 
 
-  public long getCommentCount() {
+  public Integer getCommentCount() {
     return commentCount;
   }
 
-  public void setCommentCount(long commentCount) {
+  public void setCommentCount(Integer commentCount) {
     this.commentCount = commentCount;
   }
 
 
-  public long getViewCount() {
+  public Integer getViewCount() {
     return viewCount;
   }
 
-  public void setViewCount(long viewCount) {
+  public void setViewCount(Integer viewCount) {
     this.viewCount = viewCount;
   }
 
 
-  public long getLikeCount() {
+  public Integer getLikeCount() {
     return likeCount;
   }
 
-  public void setLikeCount(long likeCount) {
+  public void setLikeCount(Integer likeCount) {
     this.likeCount = likeCount;
   }
